Updated
- CCleaner
- ConEmu
- DatabaseDotNET
- DbMigration
- Dropbox
- F.Lux
- Git
- Hashicorp-Packer
- Hashicorp-Terraform
- JetBrains-DataGrip

- Microsoft Office
- Micorsoft SQL Server Management Studio
- NodeJS
- OctopusDeploy
- RazerSynapse
- RemoteDesktopManager

+++ Dropbox/tools/ChocolateyInstall.ps1
@@ -1,11 +1,18 @@
 ﻿$arguments      = @{
-    url         = 'https://clientupdates.dropboxstatic.com/client/Dropbox%2031.4.25%20Offline%20Installer.exe'
-    checksum    = '92CC94147B1AE045CBECFD18997105F4AC8E9AF6B1AF6D239E512237C1F71FDA'
+    url         = 'https://clientupdates.dropboxstatic.com/client/Dropbox%2033.4.23%20Offline%20Installer.exe'
+    checksum    = 'F0CB456EF56C94DA8FD2BC4416CB70329133690B333E60BFD0ABADA8ACD70E7B'
     silentArgs  = '/s'
+}
+
+$isUprade = $false
+
+# If the Dropbox process exists
+if (Get-Process -Name 'Dropbox' -ErrorAction SilentlyContinue) {
+    $isUprade = $true
 }
 
 Install-Package $arguments
 
-if (Get-Process -Name 'Dropbox' -ErrorAction SilentlyContinue) {
+if (-not $isUprade -and (Get-Process -Name 'Dropbox' -ErrorAction SilentlyContinue)) {
     Stop-Process -processname 'Dropbox'
 }
